server: split images actions
