+++ server/image_status.go
@@ -1,0 +1,16 @@
+package server
+
+import (
+	"github.com/Sirupsen/logrus"
+	"golang.org/x/net/context"
+	pb "k8s.io/kubernetes/pkg/kubelet/api/v1alpha1/runtime"
+)
+
+// ImageStatus returns the status of the image.
+func (s *Server) ImageStatus(ctx context.Context, req *pb.ImageStatusRequest) (*pb.ImageStatusResponse, error) {
+	logrus.Debugf("ImageStatus: %+v", req)
+	// TODO
+	// containers/storage will take care of this by looking inside /var/lib/ocid/images
+	// and getting the image status
+	return &pb.ImageStatusResponse{}, nil
+}
